refactor delete router to use token and validate user

+++ part4/bloglist/controllers/blogs.js
@@ -41,7 +41,27 @@ blogsRouter.post('/', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).end()
+  }
+
+  if (decodedToken.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'unauthorized to delete this blog' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
+
+  // remove the blog from user.blogs
+  const user = await User.findById(decodedToken.id)
+  user.blogs = user.blogs.filter(b => b.toString() !== request.params.id)
+  await user.save()
+
   response.status(204).end()
 })
